Consolidate custom_sources test into core_integration
